Overhaul the `checkered` implementation. Fixes #7.
The SVG views don't yet use `TwistyPlayer`, so we need to use a separate implementation for each viewer type. As a further complication, we can't reach into the shadow dom of the 3D implementations to override their CSS.

It would be simple to handle the checkering ourselves (i.e. set the 3D `TwistyPlayer`s to uncheckered), but I eventually want to move as much of this library as possible into `cubing.js` itself.